TSF_GRIP_Task_2
This task involved predicting the optimum number of clusters in the Iris dataset using the K-means clustering algorithm, a fundamental technique in unsupervised machine learning. The dataset, containing measurements of iris flowers from three species, was explored and visualized to understand the data structure and distribution. After data standardization, the Elbow method was applied to determine the optimal number of clusters, which was found to be three. The K-means algorithm was then trained on the standardized data, and the resulting clusters were visualized using scatter plots. The process demonstrated effective clustering and the separability of species in the dataset.