TSF_GRIP_Task_6
This task involved creating a Decision Tree classifier to predict the class of iris flowers based on their features and visualizing the classifier graphically. Using the Iris dataset, which contains measurements of sepal length, sepal width, petal length, and petal width for three species of iris flowers, the data was first explored and visualized. The dataset was then preprocessed by dropping unnecessary columns and encoding the target variable. A Decision Tree classifier was trained on a 70-30 split of the data and achieved perfect classification accuracy on the test set. The model's performance was evaluated using metrics such as accuracy, precision, recall, and F1-score, all of which were perfect (1.00). The decision-making process of the classifier was visualized using a decision tree plot, providing a clear and interpretable structure.